+++ src/azure-cli/azure/cli/command_modules/acr/_help.py
@@ -786,9 +786,12 @@ examples:
   - name: Trigger a task run.
     text: >
         az acr task run -n MyTask -r MyRegistry
-  - name: Trigger a task run by overriding the context and file passed during Task create.
+  - name: Trigger a task run by overriding the context and file passed during Task create with a remote repository.
     text: >
         az acr task run -n MyTask -r MyRegistry -c https://github.com/Azure-Samples/acr-build-helloworld-node.git -f Dockerfile
+  - name: Trigger a task run by overriding the context and file passed during Task create with a local context.
+    text: >
+        az acr task run -n MyTask -r MyRegistry -c . -f Dockerfile
   - name: Trigger a task run by adding or overriding build arguments set during Task create.
     text: |
         az acr task run -n MyTask -r MyRegistry --arg DOCKER_CLI_BASE_IMAGE=docker:18.03.0-ce-git